+++ crates/evm/src/execution.cairo
@@ -27,8 +27,8 @@ fn execute(
     bytecode: Span<u8>,
     calldata: Span<u8>,
     value: u256,
-    gas_price: u64,
-    gas_limit: u64,
+    gas_price: u128,
+    gas_limit: u128,
 ) -> ExecutionResult {
     // Create a new root execution context.
     let call_ctx = CallContextTrait::new(
